make folders actually add and not just break everything
