+++ desktop/src/renderer/src/pages/FileList.tsx
@@ -1,3 +1,4 @@
+import { CurrencyBitcoin } from "@mui/icons-material";
 import { Box, useTheme } from "@mui/material";
 import {
 	FolderDisplayGrid,
@@ -8,11 +9,40 @@ import {
 import { useState } from "react";
 
 export const FileListPage = () => {
+	//@ts-ignore
 	const [rows, setRows] = useState<FolderRow[]>([]);
 	const theme = useTheme();
 	const onFolderDelete = (folder: string) => {
 		alert(folder); // TBD
 	};
+
+	function handleAddPathArrayToRows(inPaths) {
+		const newRows: FolderRow[] = [...rows];
+		let index = rows.length;
+		try {
+			for (let i = 0; i < inPaths.length; i++) {
+				const curFolderRow: FolderRow = {
+					id: index,
+					folder: inPaths[i],
+					schedule: "TestSchedule",
+					classification: "TestClassification",
+					file: "TestFILE",
+					opr: true,
+					startDate: null,
+					endDate: null,
+					soDate: null,
+					fdDate: null,
+					progress: 0,
+				};
+				index++;
+				newRows.push(curFolderRow);
+			}
+			setRows(newRows);
+		} catch (error) {
+			setRows(rows);
+			console.log("Hit error");
+		}
+	}
 
 	return (
 		<>
@@ -26,7 +56,7 @@ export const FileListPage = () => {
 					background: `${theme.palette.primary}`,
 				}}
 			>
-				<SelectFolderButton rows={rows} />
+				<SelectFolderButton onRowChange={handleAddPathArrayToRows} />
 				<ContinueButton />
 			</Box>
 			<Box
